feat(feedback-list): move feedback-list to App
