+++ src/App.vue
@@ -1,6 +1,7 @@
 <template>
   <v-app>
     <glisten-client />
+    <feedback-list :feedbacks="feedbacks" @changeStatus="changeStatus" @setNotes="setNotes" />
   </v-app>
 </template>
 
@@ -9,14 +10,84 @@ import GlistenClient from './components/GlistenClient.vue';
 
 import Vue from 'vue';
 import { Component } from 'vue-property-decorator';
-
+import FeedbackList from './components/FeedbackList.vue';
+import {
+  GET_ALL_FEEDBACKS,
+  SUBCRIPTION_FEEDBACKS,
+  CREATE_WHISP,
+  UPDATE_WHISP,
+} from '@/graphql/queries/whispQueries';
+import { IFeedback, FeedbackStatus } from './interfaces/feedback';
 /**
  * App component entry default
  */
 @Component({
-  components: {
-    GlistenClient,
+  components: { FeedbackList, GlistenClient },
+  apollo: {
+    feedbacks: {
+      query: GET_ALL_FEEDBACKS,
+      subscribeToMore: {
+        document: SUBCRIPTION_FEEDBACKS,
+        updateQuery: (previous, { subscriptionData }) => {
+          // If we added the whisp already don't do anything
+          // This can be caused by an update after a mutation
+          const existingFeedbackIndex = previous.feedbacks.findIndex(
+            (feedback: IFeedback) => feedback._id === subscriptionData.data.feedbackAdded._id,
+          );
+
+          if (existingFeedbackIndex >= 0) {
+            return {
+              feedbacks: Object.assign([], previous.feedbacks, {
+                [existingFeedbackIndex]: subscriptionData.data.feedbackAdded,
+              }),
+            };
+          }
+
+          return { feedbacks: [subscriptionData.data.feedbackAdded, ...previous.feedbacks] };
+        },
+      },
+    },
   },
 })
-export default class App extends Vue {}
+export default class App extends Vue {
+  private feedbacks: IFeedback[] = [];
+
+  private async changeStatus({
+    feedback,
+    status,
+  }: {
+    feedback: IFeedback;
+    status: FeedbackStatus;
+  }): Promise<void> {
+    this.$apollo.mutate({
+      mutation: UPDATE_WHISP,
+      variables: {
+        id: feedback._id,
+        whisp: {
+          ...feedback,
+          data: { ...feedback.data, status },
+        },
+      },
+    });
+  }
+
+  private async setNotes({
+    feedback,
+    notes,
+  }: {
+    feedback: IFeedback;
+    notes: string;
+  }): Promise<void> {
+    this.$apollo.mutate({
+      mutation: UPDATE_WHISP,
+      variables: {
+        id: feedback._id,
+        whisp: {
+          ...feedback,
+          data: { ...feedback.data, notes },
+        },
+      },
+    });
+  }
+}
 </script>
